requirements/base.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cssselect2==0.7.0
Django==4.1.7
django-advanced-filters==2.0.0
django-filter==23.1
djangorestframework==3.14.0
drf-yasg==1.21.5
fonttools==4.39.2
html5lib==1.1
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.0
Pillow==9.4.0
pycparser==2.21
pydyf==0.5.0
PyJWT==2.6.0
PyMySQL==1.0.3
pyphen==0.14.0
python-decouple==3.1
pytz==2023.2
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.18.4
six==1.16.0
sqlparse==0.4.3
tinycss2==1.2.1
uritemplate==4.1.1
urllib3==1.26.15
weasyprint==58.1
webencodings==0.5.1
zoomus==1.1.6
zopfli==0.2.2
